Implement sleeping
Bodies marked as sleeping are not simulated by the physics engine.

Bodies are marked as sleeping when their velocity is below the
`SleepingThreshold` for a duration indicated by `DeactivationTime`.

Bodies are woken up when they interact with active bodies,
when gravity changes or when the body's position, rotation, velocity
etc. are changed.