Allow constraint systems to solve arbitrary constraints and add helpers
Constraints now have an `entities` method that returns an entity array.
This is used in the generic `solve_constraint` system to get the constraint's bodies.

`XpbdConstraint` now takes the constraint's entity count with const generics.
This is needed for the `entities` method's return type unless
we use a `Vec` instead of an array.

`App` now has an `add_constraint` method for adding constraints.
Note that this only works for component constraints like joints,
and explicit ordering isn't supported.