Fix spatial query entity generations, fix removal detection and refactor
Some parts only used the entity index and not the generation, so
the entities couldn't be detected.

Entity removals are now detected at the end of each frame to avoid misses.

The spatial query pipeline now contains the colliders, and they're
updated on each frame or using
`SpatialQuery::update_pipeline()`.
This avoids querying for the colliders each time a spatial query
method is called.