Improve `GlobalTransform` change detection and handling for syncing
Previously, only `Transform` updates that happened before
the physics frame were detected and handled.

Now `Transform` can be changed at any moment, and it should act the
same as changing `Position`/`Rotation`.

`Transform` updates that happen during the physics frame also don't
override the `Position` update, so gravity and constraints are applied
normally even when constantly
changing the translation.

This requires some extra transform propagation and other computations,
but it shouldn't be that expensive and I believe it's worth it.