Add contact queries (#169)
Closes #168.

## Objective

It can often be useful to compute contacts and other geometric information for specific pairs of colliders instead of having to query for a ton of entities. Just being able to compute a contact pair or a time of impact for shapes A and B at specific positions can be very powerful.

This functionality is provided by Parry's methods like `contact` and `time_of_impact`, but it's not exposed in our APIs and requires dealing with Parry's own types. We should have our own functions that hide the unnecessary implementation details.

## Solution

This PR creates a `contact_query` module in `narrow_phase` that contains equivalents for Parry's geometric queries like `contact`, `contact_manifolds`, `time_of_impact` and so on.

`ContactManifold` also no longer contains the colliding entities, because it's not needed internally, and you shouldn't need to provide any entities for `contact_query::contact_manifolds`.

## Implemented queries

- [x] `contact`
- [x] `contact_manifolds`
- [x] `closest_points`
- [x] `distance`
- [x] `intersection_test`
- [x] `time_of_impact`
- [ ] (`nonlinear_time_of_impact`)

`nonlinear_time_of_impact` is probably niche enough that we can add it if/when there is a need for it.

## Thoughts

This PR has these contact queries under a separate module as standalone functions, but I can see why it could also make sense to have all geometric queries under one place, aka `SpatialQuery`. However, I feel like `SpatialQuery` would add unnecessary restrictions, since these contact queries don't have any external state and don't need access to the ECS world, and it's good to be able to use contact queries even when `SpatialQueryPlugin` is disabled. We can also make the docs a bit nicer this way since the API used for "real spatial queries" and these contact queries is inherently a bit different. This is why I prefer having them under a separate module.