Move narrow phase collision validity checks to broad phase (#179)
# Objective

Currently, the narrow phase checks if collisions are valid by checking if the colliders have compatible collision layers and if the bodies are sleeping. This shouldn't belong in the narrow phase, as all it should do is compute contacts based on `BroadCollisionPairs`.

## Solution

Move the collision validity checks to the broad phase.