Move all collision logic and plugins into new `collision` module (#184)
# Objective

The various collision modules are currently quite spread out illogically. The narrow phase, broad phase and contact reporting are all just under `plugins`, and collision data structs and contact queries are under `narrow_phase`. It would be good to have collision modules grouped more logically, and it would also be nice to have one place for a high-level overview of the engine's collision detection.

## Solution

Add a `collision` module and move the broad phase, narrow phase and contact reporting inside it. Also move contact queries and data out of the narrow phase module since they're not strictly related.

I also added module-level docs for `collision`.

## Future work

In the future, it might make sense to do a similar grouping for the integrator and solver with a module like `dynamics`. It would also be nice to get rid of the `components` and `resources` modules and move the structs into their appropriate places.