Add debug rendering for sleeping bodies (#185)
# Objective

The debug renderer doesn't display whether bodies are sleeping.

## Solution

Add a "sleeping color multiplier" that multiplies the HSLA properties of colors used for colliders, AABBs and joints when the bodies are sleeping. By default, it just makes the colors darker for sleeping bodies.

The multiplier can be configured with `PhysicsDebugConfig` and `DebugRender`.