Use swept AABBs for collider AABBs (#188)
# Objective

Collider AABBs are expanded based on the velocities of bodies. This is only done in the direction of linear velocity, but for angular velocity, AABBs are expanded uniformly in all directions. This means that e.g. rotating balls will unnecessarily have their AABBs expanded, which leads to extra overlaps.

## Solution

Use `compute_swept_aabb` instead of `compute_aabb` for computing the AABBs. The end position of the sweep is predicted using the linear and angular velocity. It takes the shape into account, so e.g. rotating balls won't be expanded, but long rotating planks will.