Child colliders (#154)
# Objective

A very common and important physics engine feature is to be able to attach multiple colliders to a body and position them relative to that body.

This is already possible via `Collider::compound`, but it's highly limiting, as you can not specify separate material properties, densities, collision layers or whether a collider is a sensor, or add other components like meshes to the separate colliders. It can also be useful to have separate collision events for the individual colliders instead of for just the shared rigid body entity.

Users expect it to be possible to have colliders be child entities that move with the parent and act as separate colliders, while still affecting the body like a normal compound collider would. It should be possible to nest the colliders in arbitrarily deep hierarchies.

## Solution

This PR makes it possible to attach colliders as child entities:

```rust
fn setup(mut commands: Commands) {
    commands
        .spawn((RigidBody::Dynamic, Collider::ball(0.5)))
        .with_children(|commands| {
            commands.spawn((
                Transform::from_xyz(3.0, 0.0, 0.0),
                Collider::ball(0.5),
            ));
        });
}
```

You should be able to nest colliders as much as you want, and they will all be positioned relative to each other using `Transform`. Each collider has its own AABB, collision events, collision layers, material properties and so on.

For knowing which colliders affect which rigid bodies, we use a `ColliderParent` component that stores the `Entity` of the rigid body. The component is automatically updated based on entity hierarchies.

For positioning the colliders correctly and computing contact positions and centers of mass relative to the rigid body, we can use an automatically updated `ColliderTransform` component that stores the local positional and rotational offset and scale relative to the rigid body (not necessarily the collider's direct parent!). To avoid precision issues and numerical drift, I try to avoid using `GlobalTransform` and instead use `Position` and `Rotation` wherever possible.

The mass properties of bodies are also automatically updated whenever any of the colliders are transformed or removed.

## Todo

- [x] Document child colliders in `Collider`
- [x] Wake up bodies when child colliders are transformed
- [x] Wake up bodies and update their mass properties when their colliders are removed
- [x] Remove `ColliderParent` and reset `ColliderOffset` when the collider's rigid body is removed
- [x] Verify that arbitrary nesting with translations and rotations works correctly

## Follow-up work

- Collider scale based on transforms #189
- Async colliders #190
- Colliders and AABBs should be debug rendering the sleeping status of the rigid body even as children
- Find a way to remove `ColliderStorageMap`
- Fix remaining contact issues with child colliders
  - Child collider collisions can sometimes be jittery, maybe the position isn't exactly right?
  - The orbiting motion of offset child colliders should be taken into account better when expanding AABBs
  - The center of mass seems to affect collisions a bit weirdly; this appears to be a constraint issue rather than a child collider issue though