Add `ColliderDensity` (#194)
# Objective

Setting the density of a collider is currently unnecessarily verbose and limiting. It requires giving the collider and density to a method that creates the mass property component:

```rust
ColliderMassProperties::new_computed(&Collider::ball(0.5), 2.5)
```

This makes it basically impossible to specify the density when the collider isn't known in advance, like for async colliders #190, and the API is just very inconvenient.

## Solution

Add a `ColliderDensity` component. The above becomes just this:

```rust
ColliderDensity(2.5)
```

Because `ColliderMassProperties` is always overwritten using this density, I also made its properties completely read-only with getter methods.

---

## Changelog

- Added `ColliderDensity`
- Added `Collider::mass_properties`
- Renamed `ColliderMassProperties::new_computed` to `ColliderMassProperties::new`
- Made the properties of `ColliderMassProperties` read-only
- Updated the documentation

## Migration Guide

```rust
// Before
let collider = Collider::ball(0.5);
commands.spawn((
    RigidBody::Dynamic,
    ColliderMassProperties::new_computed(&collider, 2.5),
    collider,
));

// After
commands.spawn((
    RigidBody::Dynamic,
    Collider::ball(0.5),
    ColliderDensity(2.5),
));
```