Warn about explosive behavior caused by overlapping at spawn (#202)
# Objective

When a dynamic body is spawned overlapping another rigid body, it can gain a lot of velocity, which causes explosive behavior. This can currently be a bit unclear for users who aren't familiar with the internals of the engine.

## Solution

Print a warning when a dynamic body is overlapping with another body at spawn. The warning is in the following format:

```rust
warn!(
    "{:?} and {:?} are overlapping at spawn, which can result in explosive behavior.",
    body1.entity, body2.entity,
);
```

## Potential future work

We could have an option to make these overlapping scenarios gently move the bodies into a non-penetrating state rather than cause a large burst of velocity.