Use `Transform` instead of `Position` and `Rotation` in examples (#201)
# Objective

Using `Transform` instead of the separate `Position` and `Rotation` components is more idiomatic and familiar for Bevy users, but currently the examples use the separate components. This also adds a lot of usage of feature-dependent math types like `Vector`, which can be a bit more confusing.

## Solution

Use `Transform` in all examples except where `Position` should be used for stability (in systems that run in the `SubstepSchedule`).