Fix system order ambiguities, and add a test for it (#230)
# Objective

Adding bevy_xpbd to a schedule should not cause it to report order ambiguities. This is essential for deterministic rollback to work.

## Solution

Add a test that tries to add PhysicsPlugins to a deterministic schedule, and update it once. If the schedule has ambiguitiy errors, it will panic.

False positives should be marked with `system.ambiguous_with`.

Order the init shapecast systems in prepare explicitly before the other systems, to resolve the current ambiguities.

---

## Changelog

- Fixed system order ambiguities