Fix rotations when center of mass is offset (#250)
# Objective

Fixes #246 and #244.

## Solution

If we consider `AccumulatedTranslation` to represent the translation of the center of mass (instead of the position),
it's only necessary to account for the center of mass offset when applying the translation. Since we already have access to the original position and translation, computing the change in `Position` based on the change in center of mass position and rotation is straightforward.

Usages of `AccumulatedTranslation` that assumed it refers to a translation applied directly to `Position` had to be updated. Fortunately, `AccumulatedTranslation` is often accessed through `RigidBodyQuery::current_position`, so not many changes were needed.

## Performance

There's no measurable performance impact when testing in my game with hundreds of child colliders and profiling with Tracy (external trace includes this PR, this trace is main branch):

![performance](https://github.com/Jondolf/bevy_xpbd/assets/31567043/99e759ab-b53b-4b01-bae6-482a15fb9b9e)

## Alternative solutions considered

An alternative approach is to apply the additional translation due a rotation every time we apply a rotation (during integration,  constraint solving) and take into account only the change in position of the center of mass in `update_lin_vel`. This was what I tried first, but I was unable to get it to work due to numerical inaccuracies that caused unexpected accelerations, or some other insidious bug I couldn't find :). Either way, this alternative would have required more changes, and for what it's worth I personally found it less clean as there was no longer a clean separation between updates to position and rotation (every rotation had to add a small change to the position).

## Changelog

- `AccumulatedTranslation` now refers to the translation of the center of mass, which may not equal the translation applied to `Position` when the center of mass is offset.
- Rotational motion with offset centers of mass now works as expected.
  - As a result, child colliders finally *feel* stable in my testing.