Scale debug rendering of center of mass dot by axis lengths (#268)
# Objective

Fixes #260.

Currently, the size of the debug-rendered yellow dot visualizing the center of mass in the middle of a body's axes is hard-coded. This is annoying and confusing when working at different scales than the assumed default scale.

## Solution

Derive the radius of the dot based on the average of the debug rendered rigid body axis lengths.