Fix rotation change check in integrator (#284)
# Objective

In #272, the integrator was changed to update rotation only if `delta.w != 0.0`. However, in some cases, x, y, or z can be non-zero even if w is zero, like with `AngularVelocity(Vec3::X)`, and in those cases the rotation would incorrectly be ignored.

## Solution

Make the condition stricter by also checking if x, y, and z are 0. Based on a quick test, this should also keep change detection working, which is what #272 was trying to address.