Adding round_cuboid (#300)
# Objective

Looking through the code it appears we have a round cuboid available from Parry but are missing the function to create it. Rounded cuboids are handy for several things though in my case it's providing a smooth tunnel edge for couched players entering tunnels and also offering slight "just missed it" jump handling.

## Solution

I reviewed the cubiod functions and SharedShapes to determine the round_cubiod pattern. Not sure if I've missed some things and happy to accept any changes or add tests and docs as needed.

## Changelog

Added
2d: `Collider::round_cuboid(x_length: Scalar, y_length: Scalar, border_radius: Scalar)`
3d: `Collider::round_cuboid(x_length: Scalar, y_length: Scalar, z_length: Scalar, border_radius: Scalar)`