Make `Collider` optional, allowing usage without Parry or Nalgebra (#327)
# Objective

#311 added support for custom collider backends. However, the default `Collider` still remains, which requires the Parry and Nalgebra dependencies which are unnecessary if your collision backend doesn't use them.

`Collider` and Parry should be optional.

## Solution

Add three feature flags: `default-collider`, `parry-f32`, and `parry-f64`. `parry-f32` is a default feature that also enables `default-collider`.

The Parry features are unfortunately necessary, because Parry is split into separate crates for f32 and f64, and we can't reuse the `f32` and `f64` feature flags, because then the dependencies wouldn't be optional.

Not having the `default-collider` feature disables `Collider`, the default collision backend using it, the `SpatialQueryPipeline`, the `SpatialQuery` system parameter, and `ShapeCaster`. The `RayCaster` is still available for now, but it doesn't actually raycast unless users implement the functionality themselves.

---

## Changelog

- Add three feature flags: `default-collider`, `parry-f32`, and `parry-f64`
- `ColliderAabb` no longer uses Parry's AABB, and it has new methods

## Migration Guide

- If using bevy_xpbd without default features, add either the `parry-f32` or `parry-f64` feature, unless you have a custom collision backend