Colliders from primitives (#326)
# Objective

Support creating colliders from the new geometric primitives in Bevy 0.13.

## Solution

Add an `IntoCollider` trait and implement it for primitives. For now, tori and conical frusta are not supported, as they aren't as trivial to implement.

Some new collider shapes are also supported: ellipses and regular polygons. These are implemented using custom shapes, as Parry doesn't have them built-in.

`Collider::ball` is now also `Collider::circle` in 2D and `Collider::sphere` in 3D. `Collider::cuboid` is `Collider::rectangle` in 2D. This makes the naming consistent with Bevy. 

---

## Migration Guide

- Replace `Collider::ball` with `Collider::circle` in 2D and `Collider::sphere` in 3D
- Replace `Collider::cuboid` with `Collider::rectangle` in 2D