various fixes in the `prepare/init_transforms` system (#360)
# Objective

The `prepare/init_transforms` system has multiple issues:
- Returns early when `!config.transform_to_position` from loop so only the first item of the query is being processed
- ~~`is_rb: Has<C>` is redundant in the presence of `Added<C>`~~
  - changed `is_rb: Has<C>` to `is_rigid_body: Has<RigidBody>`
- Where there is `Position` and/or `Rotation` and `config.transform_to_position` but no `Transform`, no new `Transform` is added
- When there is no `Position` and/or `Rotation` and  `config.transform_to_position` the `new_position`/`new_rotation` ignore the current transform and only take the global transform

## Solution

Added a test to cover this system and fixed the found issues. Also refactored out some of the deep nesting in the implementation. Note that the test does not yet cover `GlobalTransform`s and `parents` transforms.

---

## Changelog

### Fixed

Fixed various issues in the `prepare/init_transforms` system.