Make sensors not contribute to mass properties (#381)
# Objective

Similar to #355 by @yrns, but without special casing `ColliderDensity` defaults, and implemented with observers.

Sensor colliders currently contribute to the mass properties of rigid bodies. This differs from most peoples' expectations, and is also different from many existing engines such as Godot and Unity. Sensor colliders should have no impact on the physics simulation.

## Solution

- When the `Sensor` component is added to an entity, remove the collider's contribution on the rigid body's mass properties.
- When the `Sensor` component is removed from an entity, add the collider's mass properties to the rigid body's mass properties.

## Testing

There is a test to make sure that the mass properties of a rigid body are updated accordingly when the `Sensor` component is first added and then removed from a collider.

---

## Migration Guide

Colliders with the `Sensor` component no longer contribute to the mass properties of rigid bodies. You can add mass for them by adding another collider that is *not* a sensor, or by manually adding mass properties with the `MassPropertiesBundle` or its components.

Additionally, the mass properties of `Sensor` colliders are no longer updated automatically, unless the `Sensor` component is removed.

---------

Co-authored-by: Al McElrath <hello@yrns.org>