Support default layers and density for `ColliderConstructorHierarchy` (#386)
# Objective

When loading a (glTF) scene and generating colliders for it, it can be useful to be able to add all entities in that scene to some collision layer. However, `ColliderConstructorHierarchy` (previously `AsyncSceneCollider`, see #378) doesn't currently allow specifying default collision layer or density configurations despite allowing configurations for individual entities.

## Solution

Add `default_layers` and `default_density` properties for `ColliderConstructorHierarchy`. The `layers` and `density` properties of `ColliderConstructorHierarchyData` are now `Option`s, and if they are `None`, the default values are used.

I also moved the collider constructor types and tests to their own module, and updated some docs to remove mesh-specific wording.

For reviewing purposes, the first commit (8cbe48c) should be easier to read, the other commit just moves code to the `constructor` module.