Implement collision margin (#393)
# Objective

Collisions for thin objects such as triangle meshes can sometimes have stability and performance issues, especially when dynamically colliding against other thin objects.

- Infinitely thin objects are more prone to numerical issues and tunneling. Two trimeshes colliding against each other can easily get stuck in an intersecting state or have other collision issues.
- Collision algorithms for intersecting objects are typically more expensive than for the non-intersecting state. Trimesh collisions can be very expensive if the trimesh has dense geometry.

To allow users to mitigate these issues when necessary, engines such as Rapier and Bullet provide a *collision margin* or *contact skin* that forces the solver to maintain an artificial separation distance between objects. It can be thought of as a kind of shell that adds thickness to colliders, reducing the issues mentioned above.

## Solution

Add a `CollisionMargin` component. It can be used to add artifical thickness to any collider for collisions.

```rust
commands.spawn((
    RigidBody::Dynamic,
    Collider::trimesh_from_mesh(&mesh),
    CollisionMargin(0.05),
));
```

Without a collision margin:

https://github.com/Jondolf/bevy_xpbd/assets/57632562/0364168b-81e9-4887-a2fa-10a23b87481e

With a collision margin:

https://github.com/Jondolf/bevy_xpbd/assets/57632562/b41a3cbc-0709-4c3f-9f20-1b73d2d48e76

---------

Co-authored-by: Jan Hohenheim <jan@hohenheim.ch>