Add `#[reflect(Serialize, Deserialize)]` and register more types (#395)
# Objective

Updated version of #351.

A lot of types are missing `ReflectSerialize` and `ReflectDeserialize` implementations with the `serialize` feature enabled, and several types are also not registered in the `TypeRegistry`.

## Solution

Add `#[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize)]` for types that were missing it, and register more types in the `PhysicsTypeRegistrationPlugin`. I also fixed and added some other derives, including removing `Component` from `CoefficientCombine`.

---

## Migration Guide

`CoefficientCombine` is no longer a component. It was never supposed to be one, and did nothing on an entity anyways.

---------

Co-authored-by: Jan Hohenheim <jan@hohenheim.ch>