Fix solver contact normal for rotated child colliders (#409)
# Objective

Fixes #407.

The solver rework caused a regression where contact normals didn't take the rotation of child colliders into account in the contact solver, leading to buggy collisions.

## Solution

Transform the normal correctly. Also remove unnecessary transformations of the normals of the individual contact points in the contact solver, as they are not used anywhere.