Fix collider scale propagation issues (#472)
# Objective

Fixes #287 and part of #471.

Scale is currently not propagated correctly for colliders if the rigid body is not the root entity. Additionally, negative scale is prohibited for child colliders: scale is clamped to be greater than zero.

## Solution

Fix scale propagation and allow negative scale.

Note that mirroring is *not* handled properly for child colliders and some shapes yet. That will be fixed later.